fix(svc-external): fix DTOs
- Change KlineDataRequestCoinBase from abstract to concrete class.
- Introduce MexcStatusConverter for better JSON handling of trading pair statuses.
- Update MarketCapUsd property type from int? to long? for improved precision.

+++ src/SVC_Microservices/SVC_External/Services/MarketDataProviders/CoinGeckoService.cs
@@ -58,7 +58,7 @@ public class CoinGeckoService(ICoinGeckoClient coinGeckoClient) : ICoinGeckoServ
             new()
             {
                 Id = coinGeckoAssetInfo.Id,
-                MarketCapUsd = Convert.ToInt32(coinGeckoAssetInfo.MarketCapUsd),
+                MarketCapUsd = Convert.ToInt64(coinGeckoAssetInfo.MarketCapUsd),
                 PriceUsd = coinGeckoAssetInfo.PriceUsd,
                 PriceChangePercentage24h = coinGeckoAssetInfo.PriceChangePercentage24h,
                 IsStablecoin = stablecoinIds.Contains(coinGeckoAssetInfo.Id),
